# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aries Next API
  description: Aries Next 博客系统 API（由 utoipa 注解自动生成，权威契约即本文件）
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/links:
    get:
      tags:
      - Admin Links
      summary: 分页查询友情链接
      description: 稳定排序 `sort_order ASC, id ASC`，不含已软删除记录。
      operationId: listLinks
      parameters:
      - name: page
        in: path
        description: 页码，从 1 开始。
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: page_size
        in: path
        description: 每页条数（1–100）。
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: status
        in: path
        description: 按状态过滤：`active` / `inactive`。
        required: true
        schema:
          type:
          - string
          - 'null'
      - name: category_id
        in: path
        description: 按友链分类过滤。
        required: true
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: keyword
        in: path
        description: 对标题与 URL 做模糊匹配；空白值被忽略。
        required: true
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: 友情链接分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkPageResponse'
        '400':
          description: status 参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    post:
      tags:
      - Admin Links
      summary: 创建友情链接
      description: '`url` 仅允许 http/https，否则返回 400 `INVALID_LINK_URL`；`status` 缺省为 `active`。'
      operationId: createLink
      requestBody:
        description: 友情链接字段
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPayload'
        required: true
      responses:
        '201':
          description: 返回新建的友情链接
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: URL 非法或 status 非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/links/categories:
    get:
      tags:
      - Admin Links
      summary: 友链分类列表
      operationId: listLinkCategories
      responses:
        '200':
          description: 友链分类列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    post:
      tags:
      - Admin Links
      summary: 创建友链分类
      description: Slug 缺省时取 `name`，唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: createLinkCategory
      requestBody:
        description: 友链分类字段
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPayload'
        required: true
      responses:
        '201':
          description: 返回新建的友链分类
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: slug 冲突（TAXONOMY_CONFLICT）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/links/categories/{id}:
    put:
      tags:
      - Admin Links
      summary: 更新友链分类
      description: 唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: updateLinkCategory
      parameters:
      - name: id
        in: path
        description: 友链分类 ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: 友链分类字段
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPayload'
        required: true
      responses:
        '200':
          description: 返回更新后的友链分类
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 友链分类不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: slug 冲突（TAXONOMY_CONFLICT）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    delete:
      tags:
      - Admin Links
      summary: 删除友链分类
      description: 仍存在未删除的友链引用时返回 409 `TAXONOMY_IN_USE`（FK 为 `ON DELETE SET NULL`，回收站中的友链不阻塞删除）。
      operationId: deleteLinkCategory
      parameters:
      - name: id
        in: path
        description: 友链分类 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 友链分类已删除
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 友链分类不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 分类仍被引用（TAXONOMY_IN_USE）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/links/{id}:
    get:
      tags:
      - Admin Links
      summary: 获取友情链接详情
      operationId: getLink
      parameters:
      - name: id
        in: path
        description: 友情链接 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 友情链接详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 友情链接不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    put:
      tags:
      - Admin Links
      summary: 全量更新友情链接
      description: '`url` 仅允许 http/https；`status` 缺省为 `active`。'
      operationId: updateLink
      parameters:
      - name: id
        in: path
        description: 友情链接 ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: 友情链接字段
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPayload'
        required: true
      responses:
        '200':
          description: 返回更新后的友情链接
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: URL 非法或 status 非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 友情链接不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    delete:
      tags:
      - Admin Links
      summary: 软删除友情链接
      operationId: deleteLink
      parameters:
      - name: id
        in: path
        description: 友情链接 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 友情链接已删除
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无内容管理权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 友情链接不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
...
components:
  schemas:
...
    CategoryPayload:
      type: object
      description: 友链分类创建/更新请求体。
      required:
      - name
      properties:
        name:
          type: string
        slug:
          type:
          - string
          - 'null'
          description: 缺省时取 `name`。
    CategoryResponse:
      type: object
      title: CategoryResponse
      description: 管理端分类 DTO；`/api/admin/categories` 固定 `kind = article`，link / gallery 分类由各自模块管理。
      required:
      - id
      - kind
      - name
      - slug
      - description
      properties:
        description:
          type: string
        id:
          type: integer
          format: int64
        kind:
          type: string
          description: 分类类型（本模块固定为 `article`）。
        name:
          type: string
        parent_id:
          type:
          - integer
          - 'null'
          format: int64
        slug:
          type: string
...
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 机器可读错误码，如 `UNAUTHORIZED`、`ARTICLE_NOT_FOUND`。
        details:
          description: 结构化补充信息（如引用计数），大多数错误为空，序列化时省略。
        message:
          type: string
          description: 人类可读错误信息。
    ErrorResponse:
      type: object
      description: |-
        统一错误响应体（与 `http::error::ApiError` 的实际 JSON 形状一致）：
        `{ "error": { "code", "message", "details?" } }`。
        所有端点的 4xx/5xx 响应统一引用本类型。
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    LinkPageResponse:
      type: object
      required:
      - items
      - total
      - page
      - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LinkResponse'
        page:
          type: integer
          format: int32
          minimum: 0
        page_size:
          type: integer
          format: int32
          minimum: 0
        total:
          type: integer
          format: int64
    LinkPayload:
      type: object
      description: 友情链接创建/更新请求体。
      required:
      - title
      - url
      properties:
        category_id:
          type:
          - integer
          - 'null'
          format: int64
          description: 可空（允许未分类）；须指向 kind 为 `link` 的分类。
        description:
          type: string
        icon_url:
          type:
          - string
          - 'null'
          description: 图标地址，可空。
        sort_order:
          type: integer
          format: int32
        status:
          type:
          - string
          - 'null'
          description: '`active` / `inactive`，缺省 `active`。'
        title:
          type: string
          description: 标题。
        url:
          type: string
          description: 链接地址，仅允许 http/https，否则返回 400 `INVALID_LINK_URL`。
    LinkResponse:
      type: object
      description: 友情链接响应体。
      required:
      - id
      - title
      - url
      - description
      - status
      - sort_order
      - created_at
      - updated_at
      properties:
        category_id:
          type:
          - integer
          - 'null'
          format: int64
        created_at:
          type: string
          format: date-time
        description:
          type: string
        icon_url:
          type:
          - string
          - 'null'
          description: 图标地址，可空。
        id:
          type: integer
          format: int64
        sort_order:
          type: integer
          format: int32
        status:
          type: string
          description: '`active` / `inactive`。'
        title:
          type: string
          description: 标题。
        updated_at:
          type: string
          format: date-time
        url:
          type: string
          description: 链接地址，仅允许 http/https。
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aries_admin_session